Home: Add button links to other pages
This adds View Schedule, Submit Days Off, and Manage Users buttons to the main
page. This could probably be abstracted to layout.xhtml, but I'm not really
sure how best to do that yet.

Manage Users will only show up when the logged in user is the admin.

This also assumes that home.xhtml will be the place where the table of hours is
shown for to the user.