CORRECTLY do get parameters
This is nifty, I was doing get params wrong and this makes it a lot easier.

So, for reference: In the xhtml file you define what are incoming get
parameters and which bean instance variables to bind them to.

In the navigation rule, you define which attributes of the page
you're leaving to bind to new view parameters which are then passed
to the page you're navigating to. :light-bulb: