Got showSchedule working...
Just need to figure out if users or employees will be our main user
class. Currently it just finds the schedule of id 5. Change to current
id when available.

There was some weird import in SessionBean that didn't exist so i
commented it out.

I also accidentaly had selectDays as html instead of xhtml, still trying
to figure out how to get that page working with java and the db