Working on letting the user select their preferred work days
Made a new class that holds the employee's preferences since I needed a
writeable object and I cant have one in employee. So now employee holds
an EmployeePreferences object.

Added getters/setters in employee. Along with a function that will be
used in the future for setting their schedule based on the preferences
checked.

Added a button for selectDays in home

Adjusted format for selectDays