Add custom query support to DBO!!
This was a real biatch to implement because of how reflection works in java
but it works!! We can now query for arbitrary sets of an object, which is
cool. It's way faster than the old way I was mass-instantiating objects,
so I updated the employee editor to use it as well as to create a page
showing the shifts that are available.