Add skeleton methods and classes
To reflect the changes in the database --

The Room object is obviously the definition of a room.
It has a roomid and also a room number

The Booking object is a booking/stay. It represents a user's
stay at the hotel. A booking can have multiple rooms attached
to it each with their own date span. A booking can also have
charges attached to it at the discretion of the receptionist.

A Room Booking is a link between a room and a booking. It defines
which room is attached to the booking, for which nights it is
reserved, and also the price that was locked in at booking time.

A Room Price is a date-governed price override for rooms. If one
is not present for a particular room, the default price of 100 is
to be used.

A Charge is an arbitrary charge that is attached to a booking.

A Payment is a credit towards the balance of a booking. A booking
is considered paid when the sum of the payments toward that booking
equal the sum of the room costs and the charges levied against
that booking.

That's about it for now.