Add Create New Booking form
The booking creation flow is going to be like this:

You create a new booking by name. The booking is created
with the status of unconfirmed. The booking will show up on the
user's home page as being unconfirmed. The user can then click
on the booking to edit it. Editable actions are to add rooms to
the booking for certain date ranges and delete rooms from the
booking.

Once the booking is satisfactory for the user they can `confirm`
it and lock it in. Once the booking is confirmed, they can only
cancel it up until the first day of the booking.