Create BookingEditor, ReserveRoom controllers
This allows a user to view a pending booking and make changes to it.

This sets up the ReserveRoom controller and template/form with a datepicker
and such so that a user can input their desired room preferences.
TODO: Hook up logic in ReserveRoom such that eligible rooms are found
and if they are available, add them to the reservation.

TODO in BookingEditor: hook up the confirm button so that it changes
the state of the booking to confirmed.