Add frontdesk business. Correct date error!!!
tHERE WAS Aa major date parsing error that was causing room prices
to be erratic. I've fixed it. Thank goodness.

The FrontDesk controller is for the Employee dashboard, lovingly called
the Front Desk. It allows employees to edit and add charges to bookings.
Also, to check in bookings, and check them out when done.